requirements.txt
asn1crypto==1.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
cssmin==0.2.0
Flask==1.1.1
Flask-Assets==2.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
flask-mail==0.9.1
PyJWT==1.7.1
bcrypt==3.2.0
grafana-api==1.0.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pycparser==2.20
pyOpenSSL==19.1.0
pyserial==3.4
PySocks==1.7.1
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
urllib3==1.25.8
webassets==2.0
Werkzeug==1.0.0
WTForms==2.2.1
